Refactor: move nfpm fields files into subpackage
